+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     01. Default prefix is now empty.
     01. Uses the URL and Filename safe type base64 encoding scheme to encode invalid token bodies.
     01. :gift: Added validation utility methods.
+01. :gift: Added `Identity.resolve(Class)` method.
 01. :gift: Added SPI based Jackson type ID resolver.
 01. :gift: Added `orne-generators` support for identities and identity beans.